Added auth to all fetch headers

+++ src/components/workoutDetails.js
@@ -1,12 +1,22 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
 import { formatDistanceToNow } from "date-fns";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 const WorkoutDetails = ({ workout }) => {
     const  { dispatch } = useWorkoutsContext();
+    const { user } = useAuthContext();
     
     const handleClick = async () => {
+        if (!user) {
+            return
+        };
+
         const res = await fetch('/api/workouts/' + workout._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Content-Type': 'application/json',
+                'Authorization': `Bearer ${user.token}`
+            }
         });
         const json = await res.json();
